+++ dsmr_influxdb/admin.py
@@ -12,6 +12,10 @@ class InfluxdbIntegrationSettingsAdmin(SingletonModelAdmin):
         (
             None, {
                 'fields': ['enabled', 'hostname', 'port', 'secure', 'username', 'password', 'database'],
+                'description': _(
+                    'The backend process should automatically restart to apply changes. To manually restart, '
+                    'see the <a href="https://dsmr-reader.readthedocs.io/nl/v4/faq/restart_processes.html">FAQ</a>.'
+                )
             }
         ),
         (